Clean up & refactor
Refactored method to build request specific variables within
request functions, so if the user changes the params after
initialization, behavior is maintained.

* Move more info into the returned object
* use that moved info inside of the request methods

Also, there is an issue with the string parsing, where
buckets delimited with '.' will fail, because the bucket
and region cannot be located / built properly. I added
a warning in the code and in the README to publicize this.